Add files via upload
Afterthoughts:
Is it really so critical to strictly reference teams by id? Why not use team names in the games table? It would make database creation take far fewer steps since you could make the entire games table in one file read loop, and then generate a teams table more freely and securely from a simple SELECT DISTINCT on the games table. In this form, you must double the work in the read loop to scan the file for new teams and input them before you can start adding any games. Furthermore, no one's going to memorize which team is which id, so whenever you want to see the team names for a game (which will be EVERY time), you MUST do a join. 

To me this is a bad idea because the space save must be neglibible, and every time you add a new game in the future, you have to check to see if the teams exist in the teams table. The only justification I can imagine is if the database is intended to be multilingual. But it isn't -- neither the teams table name is "teams_EN" nor the teams name column is "name_EN". Furthermore, even if so, every team name is unique, 1 per country (right?), so you could just have the english name be the key/id of each team across the database, and the primary key of the teams_localized table.